requirements.txt
aniso8601==7.0.0
blinker==1.7.0
boto3==1.33.8
botocore==1.33.8
click==8.1.7
flask==3.0.0
Flask-Cors==4.0.0
Flask-GraphQL==2.0.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
promise==2.3
python-dateutil==2.8.2
Rx==1.6.3
s3transfer==0.8.2
six==1.16.0
urllib3==1.26.18
werkzeug==3.0.1
zipp==3.17.0
